move blockfetch initiator to typestate API
