+++ crates/amaru-protocols/src/connection.rs
@@ -21,8 +21,7 @@ use amaru_pure_stage::{DeserializerGuards, Effects, StageRef, Void, register_dat
 
 use crate::{
     blockfetch::{
-        self, BlockFetchMessage, Blocks, StreamBlocks, register_blockfetch_initiator,
-        register_blockfetch_initiator_pipelined, register_blockfetch_responder,
+        self, BlockFetchMessage, Blocks, StreamBlocks, register_blockfetch_initiator, register_blockfetch_responder,
     },
     chainsync::{
         self, ChainSyncInitiatorMsg, InitiatorResult, register_chainsync_initiator, register_chainsync_responder,
@@ -396,25 +395,14 @@ async fn do_handshake(
             ConnectionMessage::ChildDied(ChildId::ChainSync),
         )
         .await;
-        let blockfetch_initiator = if let Some(n) = config.blockfetch_pipeline_n {
-            register_blockfetch_initiator_pipelined(
-                &muxer,
-                peer,
-                n,
-                &eff,
-                ConnectionMessage::ChildDied(ChildId::BlockFetch),
-            )
-            .await
-        } else {
-            register_blockfetch_initiator(
-                &muxer,
-                peer,
-                *conn_id,
-                &eff,
-                ConnectionMessage::ChildDied(ChildId::BlockFetch),
-            )
-            .await
-        };
+        let blockfetch_initiator = register_blockfetch_initiator(
+            &muxer,
+            peer,
+            config.blockfetch_pipeline_n,
+            &eff,
+            ConnectionMessage::ChildDied(ChildId::BlockFetch),
+        )
+        .await;
         let peer_sharing_initiator = register_peer_sharing_initiator(
             &muxer,
             peer,
