+++ crates/amaru-protocols/src/manager/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{collections::BTreeMap, net::SocketAddr, num::NonZeroUsize, sync::Arc, time::Duration};
+use std::{collections::BTreeMap, net::SocketAddr, num::NonZeroU8, sync::Arc, time::Duration};
 
 use amaru_kernel::{EraHistory, NetworkMagic, Peer, Point};
 use amaru_observability::{Instrument, TraceContext, debug, debug_span, error, info};
@@ -268,9 +268,9 @@ pub struct ManagerConfig {
     pub accept_interval: Duration,
     pub three_strike_window: Duration,
     pub tx_submission_params: ResponderParams,
-    /// When `Some`, the BlockFetch initiator uses N lock-step typestate instances
-    /// and a fused pipeline cursor. `None` keeps the lock-step `miniprotocol` handler.
-    pub blockfetch_pipeline_n: Option<NonZeroUsize>,
+    /// BlockFetch initiator pipeline depth. `1` drives the lock-step typestate
+    /// instance; values greater than 1 wrap N instances in the CIP-0164 pipeliner.
+    pub blockfetch_pipeline_n: NonZeroU8,
 }
 
 impl ManagerConfig {
@@ -299,7 +299,7 @@ impl ManagerConfig {
         self
     }
 
-    pub fn with_blockfetch_pipeline_n(mut self, n: Option<NonZeroUsize>) -> Self {
+    pub fn with_blockfetch_pipeline_n(mut self, n: NonZeroU8) -> Self {
         self.blockfetch_pipeline_n = n;
         self
     }
@@ -314,7 +314,7 @@ impl Default for ManagerConfig {
             accept_interval: Duration::from_millis(100),
             three_strike_window: Duration::from_secs(60),
             tx_submission_params: ResponderParams::default(),
-            blockfetch_pipeline_n: None,
+            blockfetch_pipeline_n: NonZeroU8::MIN,
         }
     }
 }
